the cube is now working!  :)
The cube is now responding to my x,y,z data. For now  I have it set up
so that it simply rotates, once I get the yaw, pitch and roll data I
should be able to control its rotation more. I am reading the data from
a testbench and when it reaches the end of the file it jumps back and
starts reading from the beginning again. It still looks like a cube but
I am trying to see if we can make it look more like a sword